requirements.txt
cassandra-driver==3.9.0
kafka==1.3.3
parsedatetime==2.3
python-dateutil==2.6.0
